pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.5",
    "dnspython>=2.7.0",
    "flask>=3.1.2",
    "mysql-connector-python>=9.4.0",
    "pandas>=2.3.2",
    "paramiko>=4.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pymongo>=4.14.1",
    "pymysql>=1.1.2",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot==22.3",
    "python-whois>=0.9.5",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "streamlit>=1.49.0",
    "trafilatura>=2.0.0",
    "whois>=1.20240129.2",
]
